feat: Update JobListings component to limit the number of displayed jobs and show a "View All Jobs" button
The commit modifies the JobListings component to accept two new props: "limit" and "showButton". The "limit" prop determines the maximum number of jobs to display, and the "showButton" prop controls whether to show a button to view all jobs. The changes in App.vue update the JobListings component to pass the values for these props, limiting the displayed jobs to 3 and showing the "View All Jobs" button. This improves the user experience by providing a more focused job listing section and allowing users to easily access all available jobs.